[Refactor] Refactor the images command flagging process
